requirements.txt
fastapi
uvicorn[standard]
sqlmodel
passlib[bcrypt]
python-jose[cryptography]
reportlab
openpyxl
matplotlib
requests
python-dotenv
python-multipart
kivy
Pillow
